Add support for reading from stdin
This ended up being a bit of a rewrite because the way it was setup
previously didn't make it easy to add reading from stdin, but the setup
now is much nicer.

This also includes some further improvements in the TODO file that I can
look at later on (if I'm bored).