Step 1 - a very basic JSON validator
This _only_ validates the following cases for now:
- "{}"
- ""

This is done by simply checking the start and end Tokens to make sure
they're the appropriate brace. However, all other tokens inside are
ignored.

This does also include the boilerplate required to convert Chars to
Tokens - as well as passing the Result back up the chain.